fix commande update communes #2107
