+++ src/DataFixtures/Loader/LoadCommuneData.php
@@ -6,6 +6,7 @@ use App\Factory\CommuneFactory;
 use App\Manager\CommuneManager;
 use App\Manager\TerritoryManager;
 use App\Service\Import\CsvParser;
+use App\Utils\ImportCommune;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use Doctrine\ORM\EntityManagerInterface;
@@ -14,13 +15,6 @@ use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 
 class LoadCommuneData extends Fixture implements OrderedFixtureInterface
 {
-    // File found here: https://www.data.gouv.fr/fr/datasets/codes-postaux/
-    public const COMMUNE_LIST_CSV_PATH = '/src/DataFixtures/Files/codespostaux.csv';
-
-    public const INDEX_CSV_CODE_POSTAL = 0;
-    public const INDEX_CSV_CODE_COMMUNE = 1;
-    public const INDEX_CSV_NOM_COMMUNE = 2;
-
     private const FLUSH_COUNT = 1000;
 
     public function __construct(
@@ -39,7 +33,7 @@ class LoadCommuneData extends Fixture implements OrderedFixtureInterface
         $existingCpAndInseeCode = [];
         $territory = null;
 
-        $csvData = $this->csvParser->parse($this->params->get('kernel.project_dir').self::COMMUNE_LIST_CSV_PATH);
+        $csvData = $this->csvParser->parse($this->params->get('kernel.project_dir').ImportCommune::COMMUNE_LIST_CSV_PATH);
 
         // Start reading
         foreach ($csvData as $rowData) {
@@ -50,16 +44,16 @@ class LoadCommuneData extends Fixture implements OrderedFixtureInterface
                 continue;
             }
 
-            $itemCodeCommune = $rowData[self::INDEX_CSV_CODE_COMMUNE];
-            $itemCodePostal = $rowData[self::INDEX_CSV_CODE_POSTAL];
-            $itemNomCommune = $rowData[self::INDEX_CSV_NOM_COMMUNE];
+            $itemCodeCommune = $rowData[ImportCommune::INDEX_CSV_CODE_COMMUNE];
+            $itemCodePostal = $rowData[ImportCommune::INDEX_CSV_CODE_POSTAL];
+            $itemNomCommune = $rowData[ImportCommune::INDEX_CSV_NOM_COMMUNE];
 
             $keyCommune = $itemCodePostal.'-'.$itemCodeCommune;
             if (!empty($existingCpAndInseeCode[$keyCommune])) {
                 continue;
             }
 
-            $zipCode = self::getZipCodeByCodeCommune($itemCodeCommune);
+            $zipCode = ImportCommune::getZipCodeByCodeCommune($itemCodeCommune);
 
             if (null === $territory || $zipCode != $territory->getZip()) {
                 $territory = $this->territoryManager->findOneBy(['zip' => $zipCode]);
@@ -77,18 +71,6 @@ class LoadCommuneData extends Fixture implements OrderedFixtureInterface
 
         // Last flush for remaining communes
         $this->communeManager->flush();
-    }
-
-    public static function getZipCodeByCodeCommune($itemCodeCommune)
-    {
-        $codeCommune = $itemCodeCommune;
-        $codeCommune = str_pad($codeCommune, 5, '0', \STR_PAD_LEFT);
-        $zipCode = substr($codeCommune, 0, 2);
-        if ('97' == $zipCode) {
-            $zipCode = substr($codeCommune, 0, 3);
-        }
-
-        return $zipCode;
     }
 
     public function getOrder(): int
